Refactor to BigDecimal

+++ account-cmd/src/main/java/ua/dp/maxym/account/cmd/api/commands/DepositFundsCommand.java
@@ -4,8 +4,10 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import ua.dp.maxym.cqrs.core.commands.BaseCommand;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Data
 public class DepositFundsCommand extends BaseCommand {
-    private double amount;
+    private BigDecimal amount;
 }
