+++ account-query/src/main/java/ua/dp/maxym/account/query/domain/AccountRepository.java
@@ -2,12 +2,13 @@ package ua.dp.maxym.account.query.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface AccountRepository extends CrudRepository<BankAccount, String> {
     Optional<BankAccount> findByAccountHolder(String accountHolder);
 
-    List<BankAccount> findByBalanceGreaterThan(double balance);
-    List<BankAccount> findByBalanceLessThan(double balance);
+    List<BankAccount> findByBalanceGreaterThan(BigDecimal balance);
+    List<BankAccount> findByBalanceLessThan(BigDecimal balance);
 }
